+++ npm-api-maker/src/table/filters/filter-form.jsx
@@ -23,7 +23,7 @@ import Select from "../../inputs/select"
 import Services from "../../services.js"
 import {shapeComponent} from "set-state-compare/build/shape-component.js"
 import Text from "../../utils/text"
-import useBreakpoint from "../../use-breakpoint.js"
+import {useBreakpoint} from "responsive-breakpoints"
 import useI18n from "i18n-on-steroids/src/use-i18n.mjs"
 
 const styles = {}